+++ Integration Tests/DepotsTests/FeaturesDepotTests.cs
@@ -15,12 +15,18 @@ public class FeaturesDepotTests
     : SecurityDepotIntegrationTestsBase<Feature, FeaturesDepot> {
 
     protected override Feature EntityFactory(string Entropy) {
-        return DraftUtils.Feature();
+
+        Feature feature = DraftUtils.Feature();
+        feature.State = _storeManager.StoreEntityState().Result;
+        return feature;
     }
 
     public override async Task Update_Single_Success() {
         // Setting
-        Feature feature = await _storeManager.StoreFeature();
+        Feature feature = DraftUtils.Feature();
+        feature.State = await _storeManager.StoreEntityState();
+
+        await Store(feature);
 
         // Expectations
         string? oldDescription = feature.Description;